some changes on trainer

+++ Trainer/delete_trainer.php
@@ -1,5 +1,6 @@
 <?php   
 include 'trainer.html';
+include '../connect.php';
 ?>
 
 <html>
@@ -12,23 +13,30 @@ include 'trainer.html';
 
 <body>
     <form action="delete_trainer.php" method="get">
-        Trainer ID to Delete: <input type="text" name="trainer_id" required><br><br>
+        Trainer ID to Delete: 
+        <select name="trainer_id" id="trainer_id" required>
+        <option value="">Please Choose</option>
+        <?php $sql = "SELECT Trainer_ID FROM Trainer";
+        $result = mysqli_query($conn, $sql);
+            if ($result) {
+            while ($row = mysqli_fetch_assoc($result)) {
+                echo "<option value='{$row['Trainer_ID']}'>{$row['Trainer_ID']}</option>";
+                }}?>
+        </select><br><br>
         <input type="submit" value="Delete Trainer">
     </form>
 </body>
 </html>
 
 <?php
-include '../connect.php';
-
-if(isset($_GET['trainer_id'])) {
-    $trainer_id = $_GET['trainer_id'];
+if(isset($_GET['trainer_id']) && $_GET['trainer_id'] != "") {
+    $trainer_id =$_GET['trainer_id'];
 
     $sql = "DELETE FROM trainer WHERE Trainer_ID = '$trainer_id'";
     $result = mysqli_query($conn, $sql);
 
     if($result) {   
-    if ($conn->query($sql) ==TRUE ) {
+    if (mysqli_affected_rows($conn) > 0  ) {
         echo $trainer_id . " Trainer is deleted successfully";
     } else {
         echo "No record found for ID: " . $trainer_id;
